// lib/twitter/models/Get2UsersIdFollowingResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Twitter API v2
 * Twitter API v2 available endpoints
 *
 * The version of the OpenAPI document: 2.49
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Expansions } from './Expansions';
import {
    ExpansionsFromJSON,
    ExpansionsFromJSONTyped,
    ExpansionsToJSON,
} from './Expansions';
import type { Get2ListsIdFollowersResponseMeta } from './Get2ListsIdFollowersResponseMeta';
import {
    Get2ListsIdFollowersResponseMetaFromJSON,
    Get2ListsIdFollowersResponseMetaFromJSONTyped,
    Get2ListsIdFollowersResponseMetaToJSON,
} from './Get2ListsIdFollowersResponseMeta';
import type { Problem } from './Problem';
import {
    ProblemFromJSON,
    ProblemFromJSONTyped,
    ProblemToJSON,
} from './Problem';
import type { User } from './User';
import {
    UserFromJSON,
    UserFromJSONTyped,
    UserToJSON,
} from './User';

/**
 * 
 * @export
 * @interface Get2UsersIdFollowingResponse
 */
export interface Get2UsersIdFollowingResponse {
    /**
     * 
     * @type {Array<User>}
     * @memberof Get2UsersIdFollowingResponse
     */
    data?: Array<User>;
    /**
     * 
     * @type {Array<Problem>}
     * @memberof Get2UsersIdFollowingResponse
     */
    errors?: Array<Problem>;
    /**
     * 
     * @type {Expansions}
     * @memberof Get2UsersIdFollowingResponse
     */
    includes?: Expansions;
    /**
     * 
     * @type {Get2ListsIdFollowersResponseMeta}
     * @memberof Get2UsersIdFollowingResponse
     */
    meta?: Get2ListsIdFollowersResponseMeta;
}

/**
 * Check if a given object implements the Get2UsersIdFollowingResponse interface.
 */
export function instanceOfGet2UsersIdFollowingResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function Get2UsersIdFollowingResponseFromJSON(json: any): Get2UsersIdFollowingResponse {
    return Get2UsersIdFollowingResponseFromJSONTyped(json, false);
}

export function Get2UsersIdFollowingResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): Get2UsersIdFollowingResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'data': !exists(json, 'data') ? undefined : ((json['data'] as Array<any>).map(UserFromJSON)),
        'errors': !exists(json, 'errors') ? undefined : ((json['errors'] as Array<any>).map(ProblemFromJSON)),
        'includes': !exists(json, 'includes') ? undefined : ExpansionsFromJSON(json['includes']),
        'meta': !exists(json, 'meta') ? undefined : Get2ListsIdFollowersResponseMetaFromJSON(json['meta']),
    };
}

export function Get2UsersIdFollowingResponseToJSON(value?: Get2UsersIdFollowingResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'data': value.data === undefined ? undefined : ((value.data as Array<any>).map(UserToJSON)),
        'errors': value.errors === undefined ? undefined : ((value.errors as Array<any>).map(ProblemToJSON)),
        'includes': ExpansionsToJSON(value.includes),
        'meta': Get2ListsIdFollowersResponseMetaToJSON(value.meta),
    };
}