// lib/twitter/models/Get2UsersIdTimelinesReverseChronologicalResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Twitter API v2
 * Twitter API v2 available endpoints
 *
 * The version of the OpenAPI document: 2.49
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Expansions } from './Expansions';
import {
    ExpansionsFromJSON,
    ExpansionsFromJSONTyped,
    ExpansionsToJSON,
} from './Expansions';
import type { Get2UsersIdMentionsResponseMeta } from './Get2UsersIdMentionsResponseMeta';
import {
    Get2UsersIdMentionsResponseMetaFromJSON,
    Get2UsersIdMentionsResponseMetaFromJSONTyped,
    Get2UsersIdMentionsResponseMetaToJSON,
} from './Get2UsersIdMentionsResponseMeta';
import type { Problem } from './Problem';
import {
    ProblemFromJSON,
    ProblemFromJSONTyped,
    ProblemToJSON,
} from './Problem';
import type { Tweet } from './Tweet';
import {
    TweetFromJSON,
    TweetFromJSONTyped,
    TweetToJSON,
} from './Tweet';

/**
 * 
 * @export
 * @interface Get2UsersIdTimelinesReverseChronologicalResponse
 */
export interface Get2UsersIdTimelinesReverseChronologicalResponse {
    /**
     * 
     * @type {Array<Tweet>}
     * @memberof Get2UsersIdTimelinesReverseChronologicalResponse
     */
    data?: Array<Tweet>;
    /**
     * 
     * @type {Array<Problem>}
     * @memberof Get2UsersIdTimelinesReverseChronologicalResponse
     */
    errors?: Array<Problem>;
    /**
     * 
     * @type {Expansions}
     * @memberof Get2UsersIdTimelinesReverseChronologicalResponse
     */
    includes?: Expansions;
    /**
     * 
     * @type {Get2UsersIdMentionsResponseMeta}
     * @memberof Get2UsersIdTimelinesReverseChronologicalResponse
     */
    meta?: Get2UsersIdMentionsResponseMeta;
}

/**
 * Check if a given object implements the Get2UsersIdTimelinesReverseChronologicalResponse interface.
 */
export function instanceOfGet2UsersIdTimelinesReverseChronologicalResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function Get2UsersIdTimelinesReverseChronologicalResponseFromJSON(json: any): Get2UsersIdTimelinesReverseChronologicalResponse {
    return Get2UsersIdTimelinesReverseChronologicalResponseFromJSONTyped(json, false);
}

export function Get2UsersIdTimelinesReverseChronologicalResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): Get2UsersIdTimelinesReverseChronologicalResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'data': !exists(json, 'data') ? undefined : ((json['data'] as Array<any>).map(TweetFromJSON)),
        'errors': !exists(json, 'errors') ? undefined : ((json['errors'] as Array<any>).map(ProblemFromJSON)),
        'includes': !exists(json, 'includes') ? undefined : ExpansionsFromJSON(json['includes']),
        'meta': !exists(json, 'meta') ? undefined : Get2UsersIdMentionsResponseMetaFromJSON(json['meta']),
    };
}

export function Get2UsersIdTimelinesReverseChronologicalResponseToJSON(value?: Get2UsersIdTimelinesReverseChronologicalResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'data': value.data === undefined ? undefined : ((value.data as Array<any>).map(TweetToJSON)),
        'errors': value.errors === undefined ? undefined : ((value.errors as Array<any>).map(ProblemToJSON)),
        'includes': ExpansionsToJSON(value.includes),
        'meta': Get2UsersIdMentionsResponseMetaToJSON(value.meta),
    };
}